Move completion logic into own file
